% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_protobuf_message.R
\name{import_protobuf_messages}
\alias{import_protobuf_messages}
\title{Import Protocol Buffers Descriptors for spatialCMC}
\usage{
import_protobuf_messages()
}
\description{
This utility loads in the workspace the protocol buffers descriptors defined
in the \code{bnp-downscaling} library, via \code{RProtoBuf} package. These
descriptors can be used to deserialize the output of \code{\link{run_mcmc}}
function.
}
