% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_poisson_regression_mcmc.R
\name{pois_reg_mcmc}
\alias{pois_reg_mcmc}
\title{Run Spatial Markov Chain Monte Carlo sampler}
\usage{
pois_reg_mcmc(
  data,
  offset,
  cov_matrix,
  geometry,
  hier_params,
  mix_params,
  algo_params,
  out_dir = NULL
)
}
\arguments{
\item{data}{A numeric vector or matrix of shape (n_samples, n_dim). These are the observations on which to fit the model.}

\item{offset}{A numeric vector containing the fixed offsets used in the Poisson regression.}

\item{cov_matrix}{A matrix with n_samples rows and p columns, each row is the vector of covariates associated to i-th observation}

\item{geometry}{A \code{sfc} object of size \code{n_samples}. The spatial structure to consider.}

\item{hier_params}{A text string containing the hyperparameters of the hierarchy or a file name where the hyperparameters are stored.
A protobuf message of the corresponding type will be created and populated with the parameters.}

\item{mix_params}{A text string containing the hyperparameters of the mixing or a file name where the hyperparameters are stored.
A protobuf message of the corresponding type will be created and populated with the parameters.}

\item{algo_params}{A text string containing the hyperparameters of the algorithm or a file name where the hyperparameters are stored.}

\item{out_dir}{A string. If not \code{NULL}, is the folder where to store the output.
If \code{NULL}, a temporary directory will be created and destroyed after the sampling is finished.}
}
\value{
A list whose elements are Google Protocol Buffer Messages of type
\code{spatialcmc::PoisRegAlgorithmState}. Such object store a generic iteration of the MCMC chain.
}
\description{
In this light version, this function calls the \code{run_mcmc} executable from a subprocess via \code{\link[base]{system}} command.
}
