% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_mcmc_chain}
\alias{read_mcmc_chain}
\title{Reads an MCMC chain from a protobuf file and returns it as a list of serialized messages.}
\usage{
read_mcmc_chain(filename)
}
\arguments{
\item{filename}{The path to the protobuf file containing the MCMC chain.}
}
\value{
A list of serialized protobuf messages representing the MCMC chain.
}
\description{
Reads an MCMC chain from a protobuf file and returns it as a list of serialized messages.
}
