% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cmc.R
\name{run_cmc}
\alias{run_cmc}
\title{Run Spatial Consensus Monte Carlo sampler}
\usage{
run_cmc(
  data,
  covariates = NULL,
  geometry = NULL,
  shard_allocation,
  hier_type,
  hier_params,
  mix_type,
  mix_params,
  algo_params,
  out_dir = NULL
)
}
\arguments{
\item{data}{A numeric vector or matrix of shape (\code{n_samples}, \code{n_dim}). These are the observations on which to fit the model.}

\item{covariates}{A numeric matrix of shape (n_samples, n_covariates). These are the covariates on which to fit the model.}

\item{geometry}{A \code{sfc} object of size \code{n_samples}. This represents the geometry associated to each datum. Default value is \code{NULL}}

\item{shard_allocation}{A numeric vector of size \code{n_samples}. This represents the shard in which the corresponding datum will be assigned.}

\item{hier_type}{A text string containing the enum label of the hierarchy to use in the algorithm.}

\item{hier_params}{A text string containing the hyperparameters of the hierarchy or a file name where the hyperparameters are stored. A protobuf message of the corresponding type will be created and populated with the parameters.}

\item{mix_type}{A text string containing the enum label of the hierarchy to use in the algorithm.}

\item{mix_params}{A text string containing the hyperparameters of the mixing or a file name where the hyperparameters are stored. A protobuf message of the corresponding type will be created and populated with the parameters.}

\item{algo_params}{A text string containing the hyperparameters of the algorithm or a file name where the hyperparameters are stored.}

\item{out_dir}{A string. If not \code{NULL}, is the folder where to store the output. If \code{NULL}, a temporary directory will be created and destroyed after the sampling is finished.}
}
\value{
A list whose elements are Google Protocol Buffer Messages of type \code{bayesmix::AlgorithmState}. Such object store a generic iteration of the MCMC chain.
}
\description{
In this light version, this function calls the spatialCMC executable from a subprocess via \code{\link[base]{system}} command.
}
