% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_eppf.R
\name{sample_eppf}
\alias{sample_eppf}
\title{Sample from the EPPF of a (spatial) Product Partition Model}
\usage{
sample_eppf(
  n,
  geometry = NULL,
  mix_type,
  mix_params,
  algo_params,
  out_dir = NULL
)
}
\arguments{
\item{n}{A double representing the number of data.}

\item{geometry}{A \code{sfc} object of size \code{n}. Default value is \code{NULL}.}

\item{mix_type}{A text string containing the enum label of the mixing to use in the algorithm.}

\item{mix_params}{A text string containing the hyperparameters of the mixing or a file name where the hyperparameters are stored.
A protobuf message of the corresponding type will be created and populated with the parameters.}

\item{algo_params}{A text string containing the hyperparameters of the algorithm or a file name where the hyperparameters are stored.}

\item{out_dir}{A string. If not \code{NULL}, is the folder where to store the output.
If \code{NULL}, a temporary directory will be created and destroyed after the sampling is finished.}
}
\value{
A list whose elements are Google Protocol Buffer Messages of type
\code{bayesmix::AlgorithmState}. Such object store a generic iteration of the MCMC chain.
}
\description{
In this light version, this function calls the \code{run_mcmc} executable from a subprocess via \code{\link[base]{system}} command.
}
